Moved a little bit of SAML functional test code around to separate the test code testing itself from the test code testing implementation
