+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/filters/samlpolicy/SamlBasicValidationTest.groovy
@@ -21,7 +21,6 @@
 package features.filters.samlpolicy
 
 import framework.ReposeValveTest
-import org.opensaml.saml.saml2.core.Response
 import org.rackspace.deproxy.Deproxy
 import spock.lang.Unroll
 
@@ -40,7 +39,6 @@ import static javax.servlet.http.HttpServletResponse.SC_UNSUPPORTED_MEDIA_TYPE
 class SamlBasicValidationTest extends ReposeValveTest {
 
     static xmlSlurper = new XmlSlurper()
-    static samlUtilities = new SamlUtilities()
 
     def setupSpec() {
         reposeLogSearch.cleanLog()
@@ -170,54 +168,5 @@ class SamlBasicValidationTest extends ReposeValveTest {
         "invalid SAML"            | encodeBase64("<banana/>")
         "missing Issuer element"  | encodeBase64(samlResponse(status() >> assertion()))
         "empty Issuer element"    | encodeBase64(samlResponse({ 'saml2:Issuer'("") } >> status() >> assertion()))
-    }
-
-    // todo: fix name of test
-    @Unroll
-    def "verify the test SAML Utility can generate a valid SAML Response #testNum"() {
-        when: "we unmarshall the SAML string and try to validate the first Assertion's signature"
-        Response response = samlUtilities.unmarshallResponse(saml)
-        def isValidSignature = samlUtilities.validateSignature(response.assertions[0].signature)
-
-        then: "no exceptions were thrown (SAML was successfully unmarshalled)"
-        notThrown(Exception)
-
-        and: "the signature was valid"
-        isValidSignature
-
-        and: "the Issuer was set correctly"
-        response.getIssuer().value == "http://idp.external.com"
-
-        where:
-        [testNum, saml] << [
-                [0, samlResponse {
-                    'saml2:Issuer'("http://idp.external.com")
-                    'saml2p:Status' {
-                        'saml2p:StatusCode'(Value: "urn:oasis:names:tc:SAML:2.0:status:Success")
-                    }
-                    mkp.yieldUnescaped ASSERTION_SIGNED
-                }],
-                [1, samlResponse(issuer() >> status() >> assertion())],
-                [2, SAML_ONE_ASSERTION_SIGNED]
-        ]
-    }
-
-    // todo: fix name of test
-    @Unroll
-    def "verify the test SAML Utility validator will reject a SAML response with an invalid signature #testNum"() {
-        when: "we unmarshall the SAML string and try to validate the Assertion's signature"
-        Response response = samlUtilities.unmarshallResponse(saml)
-        def isValidSignature = samlUtilities.validateSignature(response.assertions[0].signature)
-
-        then: "the signature was not valid"
-        !isValidSignature
-
-        and: "the Issuer was set correctly"
-        response.getIssuer().value == "http://idp.external.com"
-
-        where:
-        testNum | saml
-        0       | SAML_ONE_ASSERTION_SIGNED.replace("    ", "")
-        1       | SAML_ASSERTION_INVALID_SIGNATURE
     }
 }
